+++ iconmanager.cpp
@@ -11,21 +11,28 @@ IconManager::IconManager(QObject *parent)
 
 }
 
-void IconManager::downloadImage(const QUrl &url, int id)
+void IconManager::downloadImage(QString urlStr, int id)
 {
+    if (!urlStr.contains("http://") || !urlStr.contains("https://")) {
+        urlStr.insert(0, "http://");
+    }
+    qDebug() << "Link is:" << urlStr;
+    QUrl url(urlStr);
+    if (!url.isValid()) {
+        qDebug() << "Invalid url";
+        return;
+    }
     QNetworkRequest request(url);
     auto* reply = manager->get(request);
-    connect(reply, &QNetworkReply::finished, this, [this, reply, id]{
+    QObject::connect(reply, &QNetworkReply::finished, this, [this, reply, id]{
         QString appDataLoc = QStandardPaths::writableLocation(QStandardPaths::AppDataLocation);
         QFile::remove(appDataLoc + "/images/" + QString::number(id) + ".ico");
         QPixmap pixmap;
         if (!pixmap.loadFromData(reply->readAll())) {
-            qDebug() << "not found icon";
+            qDebug() << "The return from the url is not an image";
             reply->deleteLater();
             return;
         }
-
-
         pixmap.save(appDataLoc + "/images/" + QString::number(id) + ".ico");
         qDebug() << "saved?";
         emit downloadFinished(id);
